fix tests

+++ vaadin-text-field-flow-parent/vaadin-text-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/textfield/tests/validation/NumberFieldValidationBasicPage.java
@@ -45,12 +45,12 @@ public class NumberFieldValidationBasicPage
         }));
 
         add(createInput(MIN_INPUT, "Set min", event -> {
-            int value = Integer.parseInt(event.getValue());
+            double value = Double.parseDouble(event.getValue());
             testField.setMin(value);
         }));
 
         add(createInput(MAX_INPUT, "Set max", event -> {
-            int value = Integer.parseInt(event.getValue());
+            double value = Double.parseDouble(event.getValue());
             testField.setMax(value);
         }));
